Crate potential_utf

A crate providing unvalidated string and character types.

Structs

PotentialCodePoint

A 24-bit numeric data type that is expected to be a Unicode scalar value, but is not validated as such.

PotentialUtf8

A byte slice that is expected to be a UTF-8 string but does not enforce that invariant.

PotentialUtf16